ref: rename/remove some stuff
