+++ src/Imlight.CoreLib/Auth/UserValidator.cs
@@ -47,7 +47,7 @@ using Imlight.CoreLib.WizardData.Models.Misc;
 using Imlight.CoreLib.WizardData.Models.Player;
 using static Imcodec.MessageLayer.Generated.LOGIN_7_PROTOCOL;
 
-namespace Imlight.CoreLib.AntiAmbrose;
+namespace Imlight.CoreLib.Auth;
 
 internal enum UserValidateResult {
