+++ src/Imlight.CoreLib/Game/Commands/Protocols/CommandAccountProtocol.cs
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-using Imlight.CoreLib.AntiAmbrose;
+using Imlight.CoreLib.Auth;
 using Imlight.CoreLib.WizardData;
 using Imlight.CoreLib.WizardData.Collections;
 using Imlight.CoreLib.WizardData.Models.Misc;
