+++ src/Imlight.CoreLib/WizardData/Collections/WizardCollection.cs
@@ -347,7 +347,6 @@ public static class WizardCollection {
 
     /// <summary>
     /// Updates the character's last time they clicked the "go to ___ (ex. commons)" button.
-    /// Confusingly, Kingsisle calls this the home button, while the wizard's house is called their dorm.
     /// </summary>
     /// <param name="wizard">The wizard to update the time for</param>
     public static void UpdateCharacterTimeWentHome(Wizard wizard, long time) {
